Test1/requirements.txt
streamlit==1.28.1
pdfplumber==0.9.0
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
langchain==0.0.354
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.1
rank-bm25==0.2.2
python-dotenv==1.0.0
rank-bm25==0.2.2
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.1